Add new kernel message (#4873)
