+++ news/2 Fixes/4775.md
@@ -1,0 +1,1 @@
+Popup a tip when opening a notebook for the first time.
