+++ src/client/common/utils/localize.ts
@@ -975,6 +975,11 @@ export namespace DataScience {
         'DataScience.showDataViewerFail',
         'Failed to create the Data Viewer. Check the Jupyter tab of the Output window for more info.'
     );
+
+    export const kernelTipMessage = localize(
+        'DataScience.kernelTipMessage',
+        'Tip: You can change your notebook kernel in the status bar.'
+    );
 }
 
 // Skip using vscode-nls and instead just compute our strings based on key values. Key values
